fc/standard: relocation factory + family rollout phase 1 — PRATO is the 2nd verified member (audio-ok).
fc_standard_config(sid_path): derives all 9 address fields from the PSID
load address (the vanilla player image has a FIXED internal layout —
verified empirically: 2760/4024 HVSC FC SIDs carry the canonical freq
table at exactly load+$564; 2639 share Jarre_2's full shape init=load,
play=load+6, vblank) + probes the player-variant byte (below).

Prato (Luca, load $1800) verifies FULL — play 181601/181601, trichotomy
audio-ok — via three family findings:

1. Pulse programs decoded BY REFERENCE (engine_model): n = fx2&7 for
   every instrument; the player indexes the 4-byte table blindly, so
   n>=4 reads PAST the nominal table into following data (Prato prog 7
   lands in the pattern-ptr region). Those 4 bytes are still the
   instrument's effective pulse-schedule parameters — captured by VALUE,
   keeping the rebuild's own layout free.

2. STALE-TAIL player variant — one static code byte at orig $2046 (the
   vibrato-skip JMP operand, no SMC writer): $EB (Jarre_2) = skip writes
   nothing; $DC (Prato) = jump into the vibrato WRITE TAIL ($1ADC), so
   vibrato-skipped instruments (fx1==0 / wave bit4 / bit2) write the
   STALE global work regs ($217C/$217D — the last vibrato computation,
   possibly another voice's) to their freq lo,hi EVERY frame. New
   cfg.std_vibrato_stale_tail, factory-probed. This single fix took
   Prato from 1.57% to 92.87% match.

3. Instrument-table growth: patterns select ids 0-31 ($Cx carries 5
   bits); Prato references inst 10, beyond Jarre's instr_count=10. The
   extract grows the decode to the max referenced id (records are real
   per-tune content wherever they sit); the composer sizes every
   instrument-table emission from the USF (cfg.instr_count is a floor).

Verify fixes: Songlengths fractional seconds ("0:19.813"); trichotomy
Check-A unwritten-default now includes the HOST's pre-init $D418=$0F
(psiddrv writes it BEFORE init — sid_init_report §1), so deferred-init
members (Prato's init makes ZERO SID writes, the $210E=$2C init variant)
verify correctly instead of failing on a phantom $D418 state diff.

Full regression clean: Hubbard 71/71, Companion 44/44, C64ME 15/15,
Jay_Derrett 17/17, FC 16/16 (Jarre_2 canary incl.).

NEXT (RE_NOTES + memory): the $Ex GLIDE — full RE done (cmd = dir bit0 +
speed-hi bits1-3; param = speed-lo hi-nibble + onset-threshold lo-nibble,
globals last-parsed-wins; shadow-MUTATING lo,hi direct writes between
vibrato and pulse). The current standard parser DISCARDS the param byte.
Blocks Entrail_Ranx (17376/127140) + likely several of the 9 remaining
sample failures.

+++ pipelines/future_composer/verify.py
@@ -114,7 +114,8 @@ def verify_canary(cfg: FCConfig, build_fn, root: str | None = None,
                     parts = line.rstrip().split('=', 1)[1].split()
                     for i, p in enumerate(parts):
                         m, s = p.split(':')
-                        per_sub_durations[i] = int(m) * 60 + int(s)
+                        # Songlengths entries may carry millis ("0:19.813")
+                        per_sub_durations[i] = int(m) * 60 + float(s)
                     break
 
     per_sub = {}
